switched to alt table in db

+++ src/service/movieReviews/deleteReview.service.ts
@@ -6,13 +6,25 @@ export const deleteReviewService = async (authToken: string, uuid: string) => {
   try {
     const user = await getAuth().verifyIdToken(authToken);
 
-    const db = client.db('MovieRatings').collection(user.uid);
+    const db = client.db('Reviews').collection('MovieReviews');
 
     const id = new ObjectId(uuid);
 
-    const result = await db.deleteOne({ _id: id });
+    const movie = await db.findOne({ _id: id })
+    
+    if (movie) {
+      if (movie.userId === user.uid) {
+        const result = await db.deleteOne({ _id: id });
+        
+        return result
+      } else {
+        throw new Error('Not authorized to delete this review')
+      }
+    }
 
-    return result;
+    throw new Error('Review not found')
+
+
   } catch (err) {
     throw err;
   }
